requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
mistralai>=0.0.10
numpy>=1.24.0
scikit-learn>=1.3.0
firecrawl>=0.0.1
supabase>=2.7.4
